Perform PCA on proteomics data from CPATC LUADs and matched healthy lung tissues

A composite function performs imputed PCA on CPATC LUADs proteomics data and plot the results

.plot_PCA_CPTAC_LUAD(gene_list)

Arguments

gene_list

gene names in a vector of characters

Details

This function

selects proteomics data from CPTAC_LUAD_Proteomics prepared by initialize_proteomics_data().

performs imputed PCA on combined tumors from CPTAC LUAD and and matched healthy lung tissues, by .protein_imputePCA().

plots results as biplot, scree and matrix plots by .biplot()

This function is not accessible to the user and will not show at the users' workspace. It can only be called by the exported EIF4F_PCA() function.

Side effects:

(1) PCA biplots (PCA score plot + loading plot) on screen and as pdf files: PCA score plot shows the clusters of samples based on their similarity and loading plot shows how strongly each characteristic influences a principal component.

(2) matrix plots on screen and as pdf files to show the quality of representation of the variables.

(3) scree plots on screen and as pdf files to display how much variation each principal component captures from the data.

See also

Examples

if (FALSE) {
.plot_PCA_CPTAC_LUAD(c(
 "EIF4E", "EIF4G1", "EIF4A1", "PABPC1",
 "MKNK1", "MKNK2", "EIF4EBP1"))
}