Summary of CNV statuses in bar plots

Provides the summary of CNV statuses of EIF4F genes in tumors from all TCGA cancer types combined and in tumors from individual TCGA cancer types

.plot_bargraph_CNV_TCGA(gene_list)

Arguments

gene_list

gene names in a vector of characters

Details

This function

selects CNV and sample type data from only EIF4F gene in the data frame TCGA_CNV_sampletype prepared from initialize_cnv_data().

uses the subset data .TCGA_CNV_sampletype_subset to perform the CNV status analysis of all inquired genes with the functions .CNV_all_cancer() and plot the results as a bar plot with .CNV_sum_barplot()

uses the same subset data to perform CNV status analysis for each gene across all tumors types with the function .CNV_ind_cancer() and plots the CNV results and .CNV_barplot()

It should not be used directly, only inside EIF4F_CNV_analysis() function.

Side effects:

(1) the stacked bar plots on screen and as pdf file to show the summary table of the CNV statuses of all gene_list in TCGA tumors

(2) stacked bar plots on screen and as saved pdf files to show CNV status of each gene in gene_list from an individual cancer type

See also

Examples

if (FALSE) {
plot_bargraph_CNV_TCGA(c(
  "TP53", "EIF4A1", "EIF4A2", "EIF4E",
  "EIF4E2", "EIF4E3", "MYC", "EIF3D", "EIF4EBP1", "EIF4G1", "EIF4G2",
  "EIF4G3", "PABPC1", "MKNK1", "MKNK2"
))
}