Compare the RNA ratios between EIF4F genes

This function generates box plots to compare the RNA ratios between EIF4F genes in tumors tumors vs adjacent normal tissues across TCGA cancer types, and a violin plot to compare RNA ratios in primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer types combined.

Usage

.plot_boxgraph_RNAratio_TCGA(
  gene01,
  gene02,
  gene03,
  gene04,
  gene05,
  gene06,
  gene07,
  gene08,
  gene09,
  gene10
)

Arguments

gene01

gene name as a string

gene02

gene name as a string

gene03

gene name as a string

gene04

gene name as a string

gene05

gene name as a string

gene06

gene name as a string

gene07

gene name as a string

gene08

gene name as a string

gene09

gene name as a string

gene10

gene name as a string

Details

This function

calculates RNA ratio across individual tumor types from all cancers with the function .RNAratio_calculation(),

select the ratio data with .RNAratio_selection() and make boxplot with .RNAratio_boxplot().

compares RNA ratio in primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer types combined with the function .RNAratio_tumortype() and plots with .violinplot().

It should not be used directly, only inside EIF4F_DEG_analysis() function.

side effects:

(1) box plots on screen and as pdf file to show differential RNA ratios in tumors vs NATs in TCGA cancer type

(2) violin plots on screen and as pdf file to show differential gene ratios in primary, metastatic tumors vs adjacent normal tissues from all combined TCGA cancer types

Examples

if (FALSE) {
.plot_boxgraph_RNAratio_TCGA(
  EIF4E = "EIF4E", EIF4E2 = "EIF4E2",
  EIF4E3 = "EIF4E3", EIF4EBP1 = "EIF4EBP1", EIF4G1 = "EIF4G1", EIF4G2 = "EIF4G2",
  EIF4G3 = "EIF4G3", EIF3D = "EIF3D", EIF4A1 = "EIF4A1", EIF4A2 = "EIF4A2"
)
}