Compare the expression of EIF4F genes

This function generates a summary box plot to compare the expression of all EIF4F genes across TCGA cancer types, box plots for differential gene expression in tumors tumors vs adjacent normal tissues for each gene. and a violin plot to compare differential gene expression in primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer types combined.

Usage

.plot_boxgraph_RNAseq_TCGA(gene_list)

Arguments

gene_list

gene names in a vector of characters

Details

This function

selects RNASeq of input genes and columns including sample.type, primary.disease, and primary.site,from the data frame TCGA_GTEX_RNAseq_sampletype that was prepared by the function initialize_RNAseq_data() to produce a subset data frame called .TCGA_GTEX_RNAseq_sampletype_subset

with the subset data, calls the functions .RNAseq_all_gene() to compare relative mRNA expression of all inquired genes and calls the functions .RNAseq_grouped_boxplot() plot the results as a grouped box plot.

performs differential expression analysis for each gene across all tumors types with the function .RNAseq_ind_gene() and plots with .RNAseq_boxplot().

performs differential expression analysis for each gene in primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer types combined with the function .RNAseq_tumortype() and plots with .violinplot().

It should not be used directly, only inside EIF4F_DEG_analysis() function.

Side effects:

(1) box plots on screen and as pdf file to compare relative gene expression in across all TCGA cancer types

(2) box plots on screen and as pdf file to show differential gene expression in tumors vs NATs in TCGA cancer types

(3) violin plots on screen and as pdf file to show differential gene expression in primary, metastatic tumors vs adjacent normal tissues from all combined TCGA cancer types

Examples

if (FALSE) {
.plot_boxgraph_RNAseq_TCGA(c(
  "EIF4G1", "EIF4G2", "EIF4G3",
  "PABPC1", "EIF4A1", "EIF4A2", "EIF4B", "EIF4H", "EIF4E",
  "EIF4E2", "EIF4E3", "EIF4EBP1", "EIF3D"
))
}