Read all phosphoproteomics related datasets from CPTAC LUAD

A wrapper function reads all phosphoproteomics related datasets from CPTAC LUAD.

initialize_phosphoproteomics_data()

Details

Side effects:

(1) CPTAC_LUAD_Phos: the phosphoproteomics data of CPTAC LUAD from the download data Phos.xlsx.

(2) CPTAC_LUAD_Clinic_Sampletype: a merged dataset from two data frames. It provides the annotation data about sample types (tumor or healthy tissues) and tumors stages of each sample

.CPTAC_LUAD_Clinic, the CPTAC clinical data from the download dataset S046_BI_CPTAC3_LUAD_Discovery_Cohort_Clinical_Data_r1_May2019.xlsx

.CPTAC_LUAD_sampletype, the CPTAC annotation data from the download dataset S046_BI_CPTAC3_LUAD_Discovery_Cohort_Samples_r1_May2019.xlsx

CPTAC_LUAD_Phos and CPTAC_LUAD_Clinic_Sampletype are stored as CPTAC_LUAD_Phos.csv and CPTAC_LUAD_Clinic_Sampletype.csv in ~/Documents/EIF_output/ProcessedData folder.

See also

Examples

if (FALSE) {
initialize_phosphoproteomics_data()
}