Calculate RNA ratios

A helper function calculates the RNA ratios between genes

.RNAratio_calculation(
  gene01 = "EIF4E",
  gene02 = "EIF4E2",
  gene03 = "EIF4E3",
  gene04 = "EIF4EBP1",
  gene05 = "EIF4G1",
  gene06 = "EIF4G2",
  gene07 = "EIF4G3",
  gene08 = "EIF3D",
  gene09 = "EIF4A1",
  gene10 = "EIF4A2"
)

Arguments

gene01

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene02

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene03

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene04

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene05

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene06

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene07

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene08

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene09

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

gene10

gene name, same input from .plot_boxgraph_RNAratio_TCGA()

Value

a data frame of differential RNA ratios in tumors vs adjacent normal tissues from individual TCGA cancer types

Details

This function

select the of RNAseq data of input genes from all TCGA samples, using the data frame TCGA_GTEX_RNAseq_sampletype prepared by the function initialize_RNAseq_data().

calculates the RNA ratio data in each TCGA samples including tumors and solid tissue normal samples for comparison.

It should not be used directly, only inside .plot_boxgraph_RNAratio_TCGA() function.

See also