Perform correlation analysis on RNAseq data from all tumors and healthy tissues

This function

finds correlating genes of EIF4F from TCGA_GTEX_RNAseq_sampletype RNAseq data generated by initialize_RNAseq_data().

identifies positively or negatively correlating genes from tumor or healthy samples by .EIF_correlation(),

finds the overlap of CORs for EIF4F subunits with .CORs_vennplot(),

compares the number of CORs for EIF4F subunits with .combine_CORs_summary() and plot the results in bargraph with .CORs_summary_bargraph()

compares the correlation strength of CORs from healthy tissues and tumors from merged data frame generated from .combine_CORs_list() and visualizes clustering results in heatmap with .CORs_coeff_heatmap()

retrieve the gene names from the clusters in heatmap with .get_cluster_genes()

performs the pathways enrichment analysis on the retrieved gene list and plot the results with .pathway_dotplot()

This function should not be used directly, only inside EIF4F_Corrgene_analysis() function.

Side effects:

(1) Venn diagrams on screen and as pdf file to show the overlap of EIF correlating genes identify from RNAseq of tissue_type

(2) bar graphs on screen and as pdf file to show the numbers of identified correlating genes for each EIF4F subunit identify from RNAseq of tissue_type

(3) heatmap on screen and as pdf file to show correlation strength and clustering pattern of EIF4F correlating genes

(4) dotplot on screen and as pdf file to show the enriched pathways in the genes from different clusters in the heatmap

.plot_Corr_RNAseq_TCGA_GTEX(tissue_type)

Arguments

tissue_type

all tissue/tumor types or one specific type

Examples

if (FALSE) {
.plot_Corr_RNAseq_TCGA_GTEX(x = "All")
.plot_Corr_RNAseq_TCGA_GTEX(x = "Lung")
}